requirements.txt
streamlit==1.32.0
supabase==2.3.4
pandas==2.1.4
plotly==5.18.0
folium==0.15.1
streamlit-folium==0.19.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2